% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPlot.R
\name{buildPlot}
\alias{buildPlot}
\title{Title}
\usage{
buildPlot(data, library = "ggplot2", ...)
}
\arguments{
\item{data}{data.frame}

\item{library}{character c("ggplot2","gg","highcharter","hc","plotly")}

\item{...}{extra arguments}
}
\value{
plot object
}
\description{
Title
}
\examples{
DT <- data.frame(ID=iris$Species,X=iris$Sepal.Length,Y=iris$Sepal.Width)
buildPlot(data=DT)

buildPlot(data=DT,library="highcharter")

}
