% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplot.R
\name{xplot}
\alias{xplot}
\title{Title}
\usage{
xplot(data, library = "ggplot2", ...)
}
\arguments{
\item{data}{data.frame}

\item{library}{character c("ggplot2","highcharter","plotly")}

\item{...}{extra arguments}
}
\value{
plot object
}
\description{
Title
}
\examples{
DT <- data.frame(ID=iris$Species,X=iris$Sepal.Length,Y=iris$Sepal.Width)
xplot(data=DT)

xplot(data=DT,library="highcharter")

}
