% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplot.R
\name{xplot}
\alias{xplot}
\title{Plot easy stuff}
\usage{
xplot(x, y = NULL, ..., add = FALSE, type = "p")
}
\arguments{
\item{x, y}{the \code{x} and \code{y} arguments provide the x and y
    coordinates for the plot.  Any reasonable way of defining the
    coordinates is acceptable.  See the function \code{\link{xy.coords}}
    for details.  If supplied separately, they must be of the same length.}

\item{...}{other \link[graphics]{graphical parameters} (see \code{\link[graphics]{par}} and
    section \sQuote{Details} below).}

\item{type}{1-character string giving the type of plot desired.  The
    following values are possible, for details, see \code{\link[graphics]{plot}}:
    \code{"p"} for points, \code{"l"} for lines,
    \code{"b"} for both points and lines,
    \code{"c"} for empty points joined by lines,
    \code{"o"} for overplotted points and lines,
    \code{"s"} and \code{"S"} for stair steps and
    \code{"h"} for histogram-like vertical lines.  Finally,
    \code{"n"} does not produce any points or lines.}
}
\value{
nothing, called for side effect (creating or drawing on a plot)
}
\description{
Work around core funs being bound to assumptions, we just want points, lines, rects, segments, text, when we want it. :)
}
\examples{
xplot(1:10, 10:1, type = "b")
xplot(rnorm(10), rnorm(10), type = "b", add = T)
xrect(runif(36), runif(36), add = FALSE)

}
